Export localStorage interaction functions from settingsSlice module
