+++ src/store.js
@@ -11,7 +11,7 @@ import { bossesReducer } from './features/bosses/index.js';
 import { mapsReducer } from './features/maps/index.js';
 import { achievementsReducer } from './features/achievements/index.js';
 import socketsReducer from './features/sockets/socketsSlice.js';
-import settingsReducer from './features/settings/settingsSlice.js';
+import settingsReducer from './features/settings/settingsSlice.mjs';
 
 export default configureStore({
     reducer: {
